refactor: split http route extensions
Break the HTTP endpoint registration monolith into focused route extension files by domain.

Add shared helpers for access checks and API responses while keeping the public route map unchanged.

Refs #169
